requirements.txt
amqp==5.1.1
asgiref==3.4.1
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
cachetools==4.2.4
celery==5.1.2
certifi==2022.12.7
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.14
django-celery-beat==2.2.1
django-redis==5.2.0
django-timezone-field==4.2.3
djangorestframework==3.14.0
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.3
httplib2==0.21.0
idna==3.4
importlib-metadata==4.8.3
kombu==5.1.0
packaging==21.3
prompt-toolkit==3.0.30
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.4
requests==2.27.1
rsa==4.9
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.6.0
